+++ nextjs-app/CHANGELOG.md
@@ -1,5 +1,26 @@
 # nextjs-app
 
+## 0.1.30
+
+### Patch Changes
+
+-   Updated dependencies
+    -   unframer@0.6.3
+
+## 0.1.29
+
+### Patch Changes
+
+-   Updated dependencies
+    -   unframer@0.6.2
+
+## 0.1.28
+
+### Patch Changes
+
+-   Updated dependencies
+    -   unframer@0.6.1
+
 ## 0.1.27
 
 ### Patch Changes